fix!: pass rest! assign as @rest to components (#261)
Previously, when you passed data as a `rest!` assign to a component, it
compiled it into a call that ran it throught a Map.merge with the rest
of the assigns at runtime.

This allowed you to access to data passed as `@foo` if you passed it as
`rest!: [foo: "bar"]`.

This implementation conflicted with the `Phoenix.LiveView.Engine` and
made it so LiveComponents didn't work.

This implementation also just differed from HEEx, which is probably
confusing.

Fixes #254

Co-authored-by: Steffen Deusch <steffen@deusch.me>